+++ tests/testthat/test-standardized_differences.R
@@ -69,8 +69,8 @@ if (require("testthat") && require("effectsize")) {
     x <- cohens_d(wt ~ am, data = mtcars, pooled_sd = FALSE)
     expect_equal(colnames(x)[1], "Cohens_d")
     expect_equal(x[[1]], 1.934, tolerance = 0.001)
-    expect_equal(x$CI_low, 1.098798, tolerance = 0.001)
-    expect_equal(x$CI_high, 2.833495, tolerance = 0.001)
+    expect_equal(x$CI_low, 1.075151, tolerance = 0.001)
+    expect_equal(x$CI_high, 2.772516, tolerance = 0.001)
   })
 
   test_that("hedges_g (and other bias correction things", {